Added a BlogPostList component to reduce duplicate
